requirements.txt
# DNS Routing Manager Dependencies

# Основные зависимости
PyYAML>=6.0          # Для работы с YAML конфигурацией
click>=8.0           # Для CLI интерфейса
typing-extensions>=4.0  # Для совместимости типов

# Для тестирования
pytest>=7.0          # Основной фреймворк тестирования
pytest-mock>=3.8     # Мокирование для тестов
pytest-cov>=4.0      # Покрытие кода тестами

# Для разработки
black>=22.0          # Форматирование кода
flake8>=5.0          # Линтер кода
mypy>=0.991          # Проверка типов

# Сетевые утилиты (опционально)
dnspython>=2.2       # Альтернатива dig для DNS запросов
requests>=2.28       # Для HTTP запросов (если понадобится)
